Rework "k8s create-dlcmv2" to "k8s create --lcm=dlcmv2"
